+++ testsuite/ada_lsp/completion.comments.filtering/main.adb
@@ -1,0 +1,6 @@
+procedure Main is
+   X : Integer;
+begin
+   --  This is a comment with Ada.Text_IO inside
+   X := 1;  -- inline comment Ada.
+end Main;